+++ packages/@molt/command/src/CommandParameter/processor/parameterTypes/exclusive.ts
@@ -1,5 +1,5 @@
 import type { Settings } from '../../../index.js'
-import type { Input } from '../../input.js'
+import type { Input } from '../../../ParameterInput/input.js'
 import type { Output } from '../../output.js'
 import { processEnvironment } from '../helpers/environment.js'
 import { Name } from '@molt/types'
