+++ packages/@molt/command/src/ParameterInput/input.ts
@@ -1,0 +1,5 @@
+import type { State } from '../builders/State.js'
+import type { Basic } from './basic.js'
+import type { Exclusive } from './exclusive.js'
+
+export type Input<$State extends State.Base> = Basic<$State> | Exclusive<$State>
